+++ skaha/src/main/java/org/opencadc/skaha/utils/CommonUtils.java
@@ -1,7 +1,10 @@
 package org.opencadc.skaha.utils;
 
+import ca.nrc.cadc.reg.client.LocalAuthority;
+import java.net.URI;
 import java.util.Base64;
 import java.util.Collection;
+import java.util.Set;
 
 
 public class CommonUtils {
@@ -13,13 +16,19 @@ public class CommonUtils {
         return null != str && !str.isBlank();
     }
 
-    public static String decodeBase64(String encodedString) {
-        byte[] decodedBytes = Base64.getDecoder().decode(encodedString);
-        return new String(decodedBytes).strip();
-    }
-
     public static String encodeBase64(final byte[] encodedValue) {
         final byte[] encodedBytes = Base64.getEncoder().encode(encodedValue);
         return new String(encodedBytes).strip();
+    }
+
+    /**
+     * Obtain the first configured Service URI for the given base standard ID.
+     *
+     * @param baseStandardID The URI to lookup.
+     * @return A single URI (first matching).  Never null.
+     */
+    public static URI firstLocalServiceURI(final URI baseStandardID) {
+        final Set<URI> serviceURIs = new LocalAuthority().getServiceURIs(baseStandardID);
+        return serviceURIs.stream().findFirst().orElseThrow(IllegalStateException::new);
     }
 }